+++ packages/@expo/prebuild-config/CHANGELOG.md
@@ -5,8 +5,8 @@
 ### 🛠 Breaking changes
 
 ### 🎉 New features
- 
-- [Android] Support `enableEdgeToEdge` field in app config.
+
+- [Android] Support `android.enableEdgeToEdge` field in app config. ([#35958](https://github.com/expo/expo/pull/35958) by [@behenate](https://github.com/behenate))
 
 ### 🐛 Bug fixes